# apps/web/openapi.json
{
  "components": {
    "schemas": {
      "ActivityCounts": {
        "properties": {
          "regulatory_queries": {
            "minimum": 0.0,
            "title": "Regulatory Queries",
            "type": "integer"
          },
          "transaction_screenings": {
            "minimum": 0.0,
            "title": "Transaction Screenings",
            "type": "integer"
          }
        },
        "required": [
          "regulatory_queries",
          "transaction_screenings"
        ],
        "title": "ActivityCounts",
        "type": "object"
      },
      "Answer": {
        "properties": {
          "as_of": {
            "format": "date",
            "title": "As Of",
            "type": "string"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "title": "Citations",
            "type": "array"
          },
          "synthesised": {
            "title": "Synthesised",
            "type": "boolean"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "citations",
          "synthesised",
          "as_of"
        ],
        "title": "Answer",
        "type": "object"
      },
      "Citation": {
        "properties": {
          "clause_id": {
            "title": "Clause Id",
            "type": "string"
          },
          "clause_path": {
            "title": "Clause Path",
            "type": "string"
          },
          "effective_from": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective From"
          },
          "effective_to": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective To"
          },
          "quote": {
            "title": "Quote",
            "type": "string"
          },
          "support": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Support",
            "type": "number"
          }
        },
        "required": [
          "clause_id",
          "clause_path",
          "quote",
          "support"
        ],
        "title": "Citation",
        "type": "object"
      },
      "ComplianceAssessment": {
        "properties": {
          "applicable_regulations": {
            "items": {
              "type": "string"
            },
            "title": "Applicable Regulations",
            "type": "array"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "title": "Citations",
            "type": "array"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "prompt_version": {
            "title": "Prompt Version",
            "type": "string"
          },
          "required_actions": {
            "items": {
              "type": "string"
            },
            "title": "Required Actions",
            "type": "array"
          },
          "risk_rating": {
            "$ref": "#/components/schemas/RiskRating"
          },
          "txn_id": {
            "title": "Txn Id",
            "type": "string"
          },
          "unresolved_questions": {
            "items": {
              "type": "string"
            },
            "title": "Unresolved Questions",
            "type": "array"
          }
        },
        "required": [
          "txn_id",
          "risk_rating",
          "applicable_regulations",
          "required_actions",
          "citations",
          "unresolved_questions",
          "model_version",
          "prompt_version"
        ],
        "title": "ComplianceAssessment",
        "type": "object"
      },
      "DailyActivity": {
        "properties": {
          "day": {
            "format": "date",
            "title": "Day",
            "type": "string"
          },
          "regulatory_queries": {
            "minimum": 0.0,
            "title": "Regulatory Queries",
            "type": "integer"
          },
          "transaction_screenings": {
            "minimum": 0.0,
            "title": "Transaction Screenings",
            "type": "integer"
          }
        },
        "required": [
          "day",
          "regulatory_queries",
          "transaction_screenings"
        ],
        "title": "DailyActivity",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "PostureReport": {
        "properties": {
          "activity": {
            "$ref": "#/components/schemas/ActivityCounts"
          },
          "commentary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commentary"
          },
          "commentary_generated": {
            "default": false,
            "title": "Commentary Generated",
            "type": "boolean"
          },
          "daily_activity": {
            "items": {
              "$ref": "#/components/schemas/DailyActivity"
            },
            "title": "Daily Activity",
            "type": "array"
          },
          "period": {
            "$ref": "#/components/schemas/ReportPeriod"
          },
          "previous_activity": {
            "$ref": "#/components/schemas/ActivityCounts"
          },
          "previous_period": {
            "$ref": "#/components/schemas/ReportPeriod"
          },
          "risk_distribution": {
            "items": {
              "$ref": "#/components/schemas/RiskCount"
            },
            "title": "Risk Distribution",
            "type": "array"
          },
          "unresolved_screenings": {
            "minimum": 0.0,
            "title": "Unresolved Screenings",
            "type": "integer"
          }
        },
        "required": [
          "period",
          "previous_period",
          "activity",
          "previous_activity",
          "risk_distribution",
          "unresolved_screenings",
          "daily_activity"
        ],
        "title": "PostureReport",
        "type": "object"
      },
      "QueryRequest": {
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "jurisdictions": {
            "anyOf": [
              {
                "items": {
                  "enum": [
                    "IN",
                    "EU",
                    "US",
                    "GLOBAL"
                  ],
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jurisdictions"
          },
          "question": {
            "minLength": 1,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "QueryRequest",
        "type": "object"
      },
      "ReportPeriod": {
        "properties": {
          "end": {
            "format": "date",
            "title": "End",
            "type": "string"
          },
          "start": {
            "format": "date",
            "title": "Start",
            "type": "string"
          }
        },
        "required": [
          "start",
          "end"
        ],
        "title": "ReportPeriod",
        "type": "object"
      },
      "RiskCount": {
        "properties": {
          "count": {
            "minimum": 0.0,
            "title": "Count",
            "type": "integer"
          },
          "risk_rating": {
            "$ref": "#/components/schemas/RiskRating"
          }
        },
        "required": [
          "risk_rating",
          "count"
        ],
        "title": "RiskCount",
        "type": "object"
      },
      "RiskRating": {
        "enum": [
          "low",
          "medium",
          "high",
          "blocked"
        ],
        "title": "RiskRating",
        "type": "string"
      },
      "TransactionDescriptionRequest": {
        "properties": {
          "description": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Description",
            "type": "string"
          }
        },
        "required": [
          "description"
        ],
        "title": "TransactionDescriptionRequest",
        "type": "object"
      },
      "TransactionPayload-Input": {
        "additionalProperties": true,
        "properties": {
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "appropriateness_assessed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appropriateness Assessed"
          },
          "counterparty_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Counterparty Type"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "high_risk_jurisdiction": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "High Risk Jurisdiction"
          },
          "instrument": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instrument"
          },
          "jurisdictions": {
            "anyOf": [
              {
                "items": {
                  "enum": [
                    "IN",
                    "EU",
                    "US",
                    "GLOBAL"
                  ],
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jurisdictions"
          },
          "kyc_status": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kyc Status"
          },
          "large_exposure_threshold_exceeded": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Large Exposure Threshold Exceeded"
          },
          "priority_sector_reporting_required": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority Sector Reporting Required"
          },
          "txn_id": {
            "minLength": 1,
            "title": "Txn Id",
            "type": "string"
          }
        },
        "required": [
          "txn_id"
        ],
        "title": "TransactionPayload",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "FinServGlobal Compliance API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/query": {
      "post": {
        "operationId": "query_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Answer"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Query"
      }
    },
    "/reports/posture": {
      "get": {
        "operationId": "posture_report_reports_posture_get",
        "parameters": [
          {
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostureReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Posture Report"
      }
    },
    "/screen": {
      "post": {
        "operationId": "screen_screen_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionPayload-Input"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplianceAssessment"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Screen"
      }
    },
    "/screen/from-description": {
      "post": {
        "operationId": "screen_from_description_screen_from_description_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionDescriptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplianceAssessment"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Screen From Description"
      }
    }
  }
}